[Silver III] Title: 주유소, Time: 396 ms, Memory: 41140 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,12 +8,12 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        int[] distances = Arrays.stream(br.readLine().split(" ")).mapToInt(Integer::parseInt).toArray();
-        int[] cities = Arrays.stream(br.readLine().split(" ")).mapToInt(Integer::parseInt).toArray();
+        long[] distances = Arrays.stream(br.readLine().split(" ")).mapToLong(Long::parseLong).toArray();
+        long[] cities = Arrays.stream(br.readLine().split(" ")).mapToLong(Long::parseLong).toArray();
 
-        int sumDistance = 0;
-        int minPrice = cities[0];
-        int total = 0;
+        long sumDistance = 0;
+        long minPrice = cities[0];
+        long total = 0;
         for(int i = 0; i<n-1; i++) {
             if(cities[i] < minPrice) {
                 total+=sumDistance;
